starter for Tickets, AddMatch, and AddStadium
What has been done is a bare bones starter for the mentioned 3 pages. Some important aspects are still not implemented, like the navbar and using the states from redux toolkit.
My go-to tutorial for react is net ninja's playlist, so you may find similarities with this code and the one from his playlist